feat: add concert api skeleton code

+++ domain/src/main/kotlin/com/jvmhater/moduticket/model/vo/Amount.kt
@@ -1,11 +1,11 @@
 package com.jvmhater.moduticket.model.vo
 
 @JvmInline
-value class Amount(val value: Int) {
+value class Amount(val value: Long) {
     // example method
     operator fun plus(b: Amount) = Amount(value + b.value)
 
     operator fun minus(b: Amount) = Amount(value - b.value)
 
-    operator fun compareTo(b: Amount): Int = (value - b.value)
+    operator fun compareTo(b: Amount): Int = (value - b.value).toInt()
 }
